+++ frontend/src/pages/ReviewWritingPage/components/AnswerListRecheckModal/components/ReviewCard/styles.ts
