+++ frontend/src/pages/ReviewWritingPage/components/NavigateBlockerModal/style.ts
