+++ frontend/src/pages/ReviewWritingPage/components/ProgressBar/styles.ts
