+++ frontend/src/pages/ReviewWritingPage/components/QnABox/style.ts
