+++ frontend/src/pages/ReviewWritingPage/components/SubmitCheckModal/style.ts
