+++ frontend/src/pages/index.tsx
@@ -3,6 +3,6 @@ export { default as ErrorPage } from './ErrorPage';
 export { default as HomePage } from './HomePage';
 export { default as LoadingPage } from './LoadingPage';
 export { default as ReviewListPage } from './ReviewListPage';
-export { default as ReviewWritingCardFormPage } from './ReviewWritingCardFromPage';
+export { default as ReviewWritingPage } from './ReviewWritingPage';
 export { default as ReviewWritingCompletePage } from './ReviewWritingCompletePage';
 export { default as ReviewZonePage } from './ReviewZonePage';
